// src/shared/models/MarketSale.ts
export type MarketSaleDocument = mongoose.Document & {
    fileUploadId: string;
    sheetSetting: string;
    fileSetting: string;
    "Panel": string;
    "Schedule": string;
    "Chanel": string;
    "Country": string;
    "Pack": string;
    "Molecules": string;
    "Manufacturer": string;
    "Corporation": string;
    "ATC4": string;
    "NFC123": string;
    "Product": string;
    "RX": string;
    "LC37": number;
    "LC36": number;
    "LC35": number;
    "LC34": number;
    "LC33": number;
    "LC32": number;
    "LC31": number;
    "LC30": number;
    "LC29": number;
    "LC28": number;
    "LC27": number;
    "LC26": number;
    "LC25": number;
    "LC24": number;
    "LC23": number;
    "LC22": number;
    "LC21": number;
    "LC20": number;
    "LC19": number;
    "LC18": number;
    "LC17": number;
    "LC16": number;
    "LC15": number;
    "LC14": number;
    "LC13": number;
    "LC12": number;
    "LC11": number;
    "LC10": number;
    "LC09": number;
    "LC08": number;
    "LC07": number;
    "LC06": number;
    "LC04": number;
    "LC03": number;
    "LC02": number;
    "UN37": number;
    "UN36": number;
    "UN35": number;
    "UN34": number;
    "UN33": number;
    "UN32": number;
    "UN31": number;
    "UN30": number;
    "UN29": number;
    "UN28": number;
    "UN27": number;
    "UN26": number;
    "UN25": number;
    "UN24": number;
    "UN23": number;
    "UN22": number;
    "UN21": number;
    "UN20": number;
    "UN19": number;
    "UN18": number;
    "UN17": number;
    "UN16": number;
    "UN15": number;
    "UN14": number;
    "UN13": number;
    "UN12": number;
    "UN11": number;
    "UN10": number;
    "UN09": number;
    "UN08": number;
    "UN07": number;
    "UN06": number;
    "UN05": number;
    "UN04": number;
    "UN03": number;
    "UN02": number;
    "UN01": number;
};

const marketSaleSchema = new mongoose.Schema(
    {
        fileUploadId: String,
        sheetSetting: String,
        fileSetting: String,
        "Panel": String,
        "Schedule": String,
        "Chanel": String,
        "Country": String,
        "Pack": String,
        "Molecules": String,
        "Manufacturer": String,
        "Corporation": String,
        "ATC4": String,
        "NFC123": String,
        "Product": String,
        "RX": String,
        "LC37": Schema.Types.Decimal128,
        "LC36": Schema.Types.Decimal128,
        "LC35": Schema.Types.Decimal128,
        "LC34": Schema.Types.Decimal128,
        "LC33": Schema.Types.Decimal128,
        "LC32": Schema.Types.Decimal128,
        "LC31": Schema.Types.Decimal128,
        "LC30": Schema.Types.Decimal128,
        "LC29": Schema.Types.Decimal128,
        "LC28": Schema.Types.Decimal128,
        "LC27": Schema.Types.Decimal128,
        "LC26": Schema.Types.Decimal128,
        "LC25": Schema.Types.Decimal128,
        "LC24": Schema.Types.Decimal128,
        "LC23": Schema.Types.Decimal128,
        "LC22": Schema.Types.Decimal128,
        "LC21": Schema.Types.Decimal128,
        "LC20": Schema.Types.Decimal128,
        "LC19": Schema.Types.Decimal128,
        "LC18": Schema.Types.Decimal128,
        "LC17": Schema.Types.Decimal128,
        "LC16": Schema.Types.Decimal128,
        "LC15": Schema.Types.Decimal128,
        "LC14": Schema.Types.Decimal128,
        "LC13": Schema.Types.Decimal128,
        "LC12": Schema.Types.Decimal128,
        "LC11": Schema.Types.Decimal128,
        "LC10": Schema.Types.Decimal128,
        "LC09": Schema.Types.Decimal128,
        "LC08": Schema.Types.Decimal128,
        "LC07": Schema.Types.Decimal128,
        "LC06": Schema.Types.Decimal128,
        "LC04": Schema.Types.Decimal128,
        "LC03": Schema.Types.Decimal128,
        "LC02": Schema.Types.Decimal128,
        "UN37": Schema.Types.Decimal128,
        "UN36": Schema.Types.Decimal128,
        "UN35": Schema.Types.Decimal128,
        "UN34": Schema.Types.Decimal128,
        "UN33": Schema.Types.Decimal128,
        "UN32": Schema.Types.Decimal128,
        "UN31": Schema.Types.Decimal128,
        "UN30": Schema.Types.Decimal128,
        "UN29": Schema.Types.Decimal128,
        "UN28": Schema.Types.Decimal128,
        "UN27": Schema.Types.Decimal128,
        "UN26": Schema.Types.Decimal128,
        "UN25": Schema.Types.Decimal128,
        "UN24": Schema.Types.Decimal128,
        "UN23": Schema.Types.Decimal128,
        "UN22": Schema.Types.Decimal128,
        "UN21": Schema.Types.Decimal128,
        "UN20": Schema.Types.Decimal128,
        "UN19": Schema.Types.Decimal128,
        "UN18": Schema.Types.Decimal128,
        "UN17": Schema.Types.Decimal128,
        "UN16": Schema.Types.Decimal128,
        "UN15": Schema.Types.Decimal128,
        "UN14": Schema.Types.Decimal128,
        "UN13": Schema.Types.Decimal128,
        "UN12": Schema.Types.Decimal128,
        "UN11": Schema.Types.Decimal128,
        "UN10": Schema.Types.Decimal128,
        "UN09": Schema.Types.Decimal128,
        "UN08": Schema.Types.Decimal128,
        "UN07": Schema.Types.Decimal128,
        "UN06": Schema.Types.Decimal128,
        "UN05": Schema.Types.Decimal128,
        "UN04": Schema.Types.Decimal128,
        "UN03": Schema.Types.Decimal128,
        "UN02": Schema.Types.Decimal128,
        "UN01": Schema.Types.Decimal128,
    }, { timestamps: true, collection: "marketSales", strict: false }
);

export const MarketSale = mongoose.model<MarketSaleDocument>("MarketSale", marketSaleSchema);